requirements.txt
# core
Django==1.11.0
gunicorn==19.7.1
django-cors-headers==2.0.2

# db
psycopg2==2.7.1
dj-database-url==0.4.2
django-model-utils==3.0.0

# threads, mq and events
gevent==1.2.1
celery[redis]==4.0.2

# api
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
markdown==2.6.8
django-filter==1.0.4

# utils
PyJWT==1.5.0

# env
ipython==6.0